+++ ApplicationCore/ApplicationCore/Features/CADCode/Contracts/Machining/Pocket.cs
@@ -6,8 +6,8 @@ public record Pocket : Token {
     public Point PositionB { get; init; } = new(0, 0);
     public Point PositionC { get; init; } = new(0, 0);
     public Point PositionD { get; init; } = new(0, 0);
-    public float StartDepth { get; init; }
-    public float EndDepth { get; init; }
+    public double StartDepth { get; init; }
+    public double EndDepth { get; init; }
     public int Climb { get; init; }
 
 }
