+++ ApplicationCore/ApplicationCore/Features/CADCode/Contracts/Machining/RouteOffset.cs
@@ -1,3 +1,3 @@
 ﻿namespace ApplicationCore.Features.CADCode.Contracts.Machining;
 
-public record RouteOffset(OffsetType Type, float Amount);
+public record RouteOffset(OffsetType Type, double Amount);
