+++ ApplicationCore/ApplicationCore/Features/CADCode/Contracts/Machining/Tool.cs
@@ -1,3 +1,3 @@
 ﻿namespace ApplicationCore.Features.CADCode.Contracts.Machining;
 
-public record Tool (string Name, float Diameter);
+public record Tool (string Name, double Diameter);
