+++ ApplicationCore/ApplicationCore/Features/CADCode/Domain/DrillBit.cs
@@ -1,3 +1,3 @@
 ﻿namespace ApplicationCore.Features.CADCode.Services.Domain;
 
-public record DrillBit(string Name, float Diameter, float SpindleSpeed, float FeedSpeed);
+public record DrillBit(string Name, double Diameter, double SpindleSpeed, double FeedSpeed);
